requirements.txt
apns2==0.7.2
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.20.37
botocore==1.23.37
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
dj-database-url==0.5.0
Django==3.2.10
django-appconf==1.0.5
django-compressor==3.1
django-crispy-forms==1.13.0
django-currentuser==0.5.3
django-environ==0.8.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-on-heroku==1.1.2
django-storages==1.12.3
django-webpush==0.3.4
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
install==1.3.5
jmespath==0.10.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pushbullet.py==0.12.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.5
python-magic==0.4.25
pytz==2021.3
rcssmin==1.1.0
requests==2.27.1
rjsmin==1.2.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
urllib3==1.26.8
websocket-client==1.2.3
whitenoise==5.3.0
